postgres

+++ backend/migrations/001_create_services.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS services (
     name          TEXT NOT NULL,
     service_type  TEXT NOT NULL,
     config        TEXT NOT NULL,
-    interval_secs INTEGER NOT NULL DEFAULT 60,
-    enabled       INTEGER NOT NULL DEFAULT 1,
+    interval_secs BIGINT NOT NULL DEFAULT 60,
+    enabled       BIGINT NOT NULL DEFAULT 1,
     created_at    TEXT NOT NULL,
     updated_at    TEXT NOT NULL
 );
